Created JSONToWorldObjects class, moved JSON reader functions from World
